Redis connection refused so far

+++ controller-scala/src/main/scala/Mixins/RedisMixin.scala
@@ -1,6 +1,6 @@
 import redis.clients.jedis.Jedis
 
 trait RedisMixin {
-  final val redis = new Jedis("localhost", System.getenv("REDIS_PORT"))
+  final val redis = new Jedis("localhost", System.getenv("REDIS_PORT").toInt)
   redis.auth(System.getenv("REDIS_PASSWORD"))
 }
